droneinit/requirements.txt
djitellopy
ping